Moved config from internal to root level
