+++ audit/audit.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"strings"
 
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/internal/utils"
 )
