+++ cmd/cli_flags/flags.go
@@ -5,7 +5,7 @@ import (
 	"log"
 	"os"
 
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/internal/utils"
 )
