+++ cmd/main.go
@@ -11,7 +11,7 @@ import (
 	"github.com/citihub/probr"
 	"github.com/citihub/probr/audit"
 	"github.com/citihub/probr/cmd/cli_flags"
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 )
 
 func main() {
