+++ config/logfilter_test.go
