+++ service_packs/kubernetes/container_registry_access/helpers.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 
 	"github.com/citihub/probr/audit"
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 	"github.com/cucumber/godog"
