+++ service_packs/kubernetes/internet_access/helpers.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"github.com/citihub/probr/audit"
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
 	"github.com/cucumber/godog"
