+++ service_packs/kubernetes/pack/kubernetes_pack_test.go
@@ -3,7 +3,7 @@ package kubernetes_pack
 import (
 	"testing"
 
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/service_packs/coreengine"
 )
