+++ service_packs/kubernetes/pack/pack.go
@@ -1,7 +1,7 @@
 package kubernetes_pack
 
 import (
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes/container_registry_access"
 	"github.com/citihub/probr/service_packs/kubernetes/general"
