+++ service_packs/kubernetes/pod_security_policy/helpers.go
@@ -8,7 +8,7 @@ import (
 	"time"
 
 	"github.com/citihub/probr/audit"
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/internal/utils"
 	"github.com/citihub/probr/service_packs/coreengine"
 	"github.com/citihub/probr/service_packs/kubernetes"
