+++ service_packs/storage/pack/pack.go
@@ -1,7 +1,7 @@
 package storage_pack
 
 import (
-	"github.com/citihub/probr/internal/config"
+	"github.com/citihub/probr/config"
 	"github.com/citihub/probr/service_packs/coreengine"
 	azure_access_whitelisting "github.com/citihub/probr/service_packs/storage/azure/access_whitelisting"
 	azure_encryption_at_rest "github.com/citihub/probr/service_packs/storage/azure/encryption_at_rest"
